Moved queryLayerViewAgeStats to a separate module.
