+++ index.html
@@ -62,8 +62,7 @@
 <script src="http://js.arcgis.com/4.10/"></script>
 
 <!--reference the javascript for statistics spatial query-->
-<script src="querttest.js"></script>
-<script src="query.js"></script>
+<script src="index.js"></script>
 
 
 
@@ -176,7 +175,7 @@
     </div>
 
 
-  <div id="viewDiv1" data-dojo-type="dijit/layout/ContentPane" data-dojo-props="region: 'center'"></div>
+  <!-- <div id="viewDiv1" data-dojo-type="dijit/layout/ContentPane" data-dojo-props="region: 'center'"></div> -->
   <div id="viewDiv" data-dojo-type="dijit/layout/ContentPane" data-dojo-props="region: 'center'"> </div>
