+++ index.js
@@ -16,7 +16,8 @@ require([
   "esri/tasks/QueryTask",
   "esri/tasks/support/Query",
   "esri/symbols/SimpleMarkerSymbol",
-  "esri/core/watchUtils"
+  "esri/core/watchUtils",
+  "./QueryLib.js"
 ], function(
   SketchViewModel,
   Polyline,
@@ -35,7 +36,8 @@ require([
   QueryTask,
   Query,
   SimpleMarkerSymbol,
-  watchUtils
+  watchUtils,
+  QueryLib
 ) {
   // App 'globals'
   let sketchViewModel, featureLayerView, pausableWatchHandle, chartExpand;
@@ -76,16 +78,6 @@ require([
       maxScale: 0,
       minScale: 77580000
     }
-  });
-
-  // Set up statistics definition for client-side query
-  // Total popultion of age groups by gender in census tracts
-  const statDefinitions = ["Hosp_beds", "Occupency"].map(function(fieldName) {
-    return {
-      onStatisticField: fieldName,
-      outStatisticFieldName: fieldName + "_TOTAL",
-      statisticType: "sum"
-    };
   });
 
   // Update UI
@@ -255,7 +247,7 @@ require([
     bufferGraphic.geometry = buffer;
 
     // Query total beds and beds occupied
-    queryLayerViewAgeStats(buffer).then(function(newData) {
+    QueryLib.queryLayerViewAgeStats(featureLayerView, buffer).then(function(newData) {
       // Create a chart from the returned result
       updateChart(newData);
     });
@@ -274,40 +266,6 @@ require([
         family: "sans-serif"
       }
     };
-  }
-
-  //Spatial query for hospital feature layer view for statistics
-  function queryLayerViewAgeStats(buffer) {
-    // Data storage for the chart
-    let Hosp_beds = [],
-      Occupency = [];
-
-    // Client-side spatial query:
-    // Get the total bed count and beds occupied for the hospitals
-    const query = featureLayerView.layer.createQuery();
-    query.outStatistics = statDefinitions;
-    query.geometry = buffer;
-
-    // Query the features on the client using FeatureLayerView.queryFeatures
-    return featureLayerView
-      .queryFeatures(query)
-      .then(function(results) {
-        // Statistics query returns a feature with 'stats' as attributes
-        const attributes = results.features[0].attributes;
-        // Loop through attributes and save the values for use in the chart.
-        for (var key in attributes) {
-          if (key.includes("Hosp")) {
-            Hosp_beds.push(attributes[key]);
-          } else {
-            Occupency.push(-Math.abs(attributes[key]));
-          }
-        }
-
-        return [Hosp_beds, Occupency];
-      })
-      .catch(function(error) {
-        console.log(error);
-      });
   }
 
   // Draw the buffer polygon when application loads or when user searches for a new location
